+++ src/elements/mod.rs
@@ -1,7 +1,9 @@
+mod b_block;
 mod b_box;
 mod b_button;
 mod b_notification;
 
+pub use b_block::*;
 pub use b_box::*;
 pub use b_button::*;
 pub use b_notification::*;