+++ website/src/examples/basic_block.rs
@@ -1,0 +1,7 @@
+use leptos::*;
+use leptos_bulma::elements::BBlock;
+
+#[component]
+pub fn BasicBlock() -> impl IntoView {
+    view! { <BBlock class="has-text-centered">"Hello, World!"</BBlock> }
+}