+++ website/src/examples/basic_box.rs
@@ -3,5 +3,5 @@ use leptos_bulma::elements::BBox;
 
 #[component]
 pub fn BasicBox() -> impl IntoView {
-    view! { <BBox class="has-text-centered m-4">"Hello, World!"</BBox> }
+    view! { <BBox class="has-text-centered">"Hello, World!"</BBox> }
 }