+++ website/src/examples/mod.rs
@@ -5,6 +5,7 @@ use web_sys::{Request, RequestInit, Response};
 
 use crate::components::CodeBlock;
 
+mod basic_block;
 mod basic_box;
 mod basic_columns;
 mod basic_dropdown;
@@ -13,6 +14,7 @@ mod basic_navbar;
 mod basic_pagination;
 mod general_form;
 
+pub use basic_block::BasicBlock;
 pub use basic_box::BasicBox;
 pub use basic_columns::BasicColumns;
 pub use basic_dropdown::BasicDropdown;