+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Checkbox Leptos component.
+- Block Leptos component.
 
 ## [0.0.0-alpha.9] - 2024-04-10